requirements.txt
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
eval_type_backport==0.2.2
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.30.2
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
sentry-sdk==2.28.0
setproctitle==1.3.6
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
stringzilla==3.12.5
torch==1.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wandb==0.19.11
zipp==3.21.0
